websearcher/requirements.txt
openai>=1.0.0
PyYAML>=6.0
tqdm>=4.66.0
requests>=2.31.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
pdfplumber>=0.10.0
pandas>=2.0.0
numpy>=1.24.0
sympy>=1.12
regex>=2023.12.25
python-dateutil>=2.8.2

# Optional (quality-of-life / better tokenization & sentence splitting)
tiktoken>=0.5.0
nltk>=3.8.1

crawl4ai

